Backend/requirements.txt
fastapi==0.112.0
uvicorn
pydantic
pymongo
passlib
bson
email-validator
python-dotenv
a2wsgi
